chore: split defaults
