+++ test/importers/defaults/transformDOM.spec.js
@@ -1,0 +1,27 @@
+/*
+ * Copyright 2023 Adobe. All rights reserved.
+ * This file is licensed to you under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License. You may obtain a copy
+ * of the License at http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under
+ * the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR REPRESENTATIONS
+ * OF ANY KIND, either express or implied. See the License for the specific language
+ * governing permissions and limitations under the License.
+ */
+
+import { strictEqual } from 'assert';
+import { describe, it } from 'mocha';
+
+import defaultTransformDOM from '../../../src/importer/defaults/transformDOM.js';
+import TestUtils from '../../TestUtils.js';
+
+const { createDocumentFromString } = TestUtils;
+
+describe('defaultTransformDOM tests', () => {
+  it('default transformation', async () => {
+    const document = createDocumentFromString('<html><body><h1>Hello World</h1></body></html>');
+    const out = await defaultTransformDOM({ document });
+    strictEqual(out.outerHTML, '<body><h1>Hello World</h1></body>');
+  });
+});
